Implement shipping_list_methods_for_checkout using new approach
