+++ saleor/graphql/webhook/resolvers.py
@@ -1,12 +1,6 @@
-from django.conf import settings
 from django.db.models import Exists, OuterRef, Q
 
 from ...app.models import App
-from ...checkout.fetch import (
-    fetch_checkout_info,
-    fetch_checkout_lines,
-    get_all_shipping_methods_list,
-)
 from ...core.exceptions import PermissionDenied
 from ...permission.enums import AppPermission
 from ...webhook import models, payloads
@@ -60,29 +54,3 @@ def resolve_sample_payload(info: ResolveInfo, event_name, app):
         if user and user.has_perm(required_permission):
             return payloads.generate_sample_payload(event_name)
         raise PermissionDenied(permissions=[required_permission])
-
-
-def resolve_shipping_methods_for_checkout(
-    info: ResolveInfo,
-    checkout,
-    manager,
-    database_connection_name: str = settings.DATABASE_CONNECTION_DEFAULT_NAME,
-):
-    lines, _ = fetch_checkout_lines(checkout)
-    shipping_channel_listings = checkout.channel.shipping_method_listings.all()
-    checkout_info = fetch_checkout_info(
-        checkout,
-        lines,
-        manager,
-        shipping_channel_listings,
-        database_connection_name=database_connection_name,
-    )
-    all_shipping_methods = get_all_shipping_methods_list(
-        checkout_info,
-        checkout.shipping_address,
-        lines,
-        shipping_channel_listings,
-        manager,
-        database_connection_name=database_connection_name,
-    )
-    return all_shipping_methods
