+++ saleor/webhook/base.py
@@ -2,12 +2,10 @@ from collections.abc import Iterable
 from typing import Any, Callable, Optional, TypeVar
 
 from django.conf import settings
-from django.utils.module_loading import import_string
 
 from ..core.middleware import Requestor
 from ..core.utils.events import call_event
 from ..permission.enums import BasePermissionEnum
-from ..plugins.manager import get_plugins_manager
 from .models import Webhook
 
 
@@ -19,6 +17,15 @@ class WebhookBase:
     permission: BasePermissionEnum
     subscription_type: str
 
+
+class SyncWebhookBase(WebhookBase):
+    # For now this class is only use to distinguish between sync and async webhooks.
+    # When migrating more webhooks to the new approach, this class could implement
+    # generic function to trigger sync webhooks.
+    pass
+
+
+class AsyncWebhookBase(WebhookBase):
     @classmethod
     def trigger_webhook_async(
         cls,
@@ -28,6 +35,7 @@ class WebhookBase:
         allow_replica: bool = True,
         legacy_data_generator: Optional[Callable] = None,
     ):
+        from ..plugins.manager import get_plugins_manager
         from .transport.asynchronous.transport import trigger_webhooks_async
         from .utils import get_webhooks_for_event
 
@@ -43,23 +51,16 @@ class WebhookBase:
                 webhooks = get_webhooks_for_event(cls)
             call_event(
                 trigger_webhooks_async,
-                None,
-                cls,
-                webhooks,
-                subscribable_object,
-                requestor,
+                data=None,
+                event_type=cls,
+                webhooks=webhooks,
+                subscribable_object=subscribable_object,
+                requestor=requestor,
                 legacy_data_generator=legacy_data_generator,
                 allow_replica=allow_replica,
             )
 
 
 WebhookBaseType = TypeVar("WebhookBaseType", bound=WebhookBase)
-
-
-def register(webhook_spec: type[WebhookBaseType]):
-    from ..graphql.webhook.subscription_types import WEBHOOK_TYPES_MAP
-
-    # Register webhook subscription type
-    WEBHOOK_TYPES_MAP[webhook_spec.event_type] = import_string(
-        webhook_spec.subscription_type
-    )
+AsyncWebhookBaseType = TypeVar("AsyncWebhookBaseType", bound=AsyncWebhookBase)
+SyncWebhookBaseType = TypeVar("SyncWebhookBaseType", bound=SyncWebhookBase)
